+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Zoom to multiple units (z keyboard shortcut).
 - Apply actions on multiple units at once (change symbol, duplicate etc.).
 - Add button for changing unit symbol directly.
 - Add undo and redo buttons to navigation bar.